--- a/test/runtests.jl
+++ b/test/runtests.jl
@@ -23,6 +23,7 @@ Random.seed!(1234)
     include("test_ordered.jl")
     include("test_contains.jl")
     include("test_probe.jl")
+    include("test_newprograms.jl")
 
     # Excluded because it contains long tests
     # include("test_realistic_searches.jl")