I added some optimizations to the `new_programs` function. The optimizations are
1. Use `SumIterator` which is an iterator I created that creates lists of values that sum to a given sum.
2. Use `NewPrograms` iterator that instead of _pushing_ into the array it consumes each new program at a time. To implement this I had to keep track of all 3 for loop that the algorithm requires and make sure that I preserve the same semantics as in the old implementation.

## Other changes
1. I ran the Julia formatter on `probe_iteartor.jl` that's why you see some small changes related to spaces 
2. I created a benchmark file in the `probe` folder to test the performance.

This is a sample output of the benchmark file on my PC
```jl
Optimized
  1.814683 seconds (12.90 M allocations: 1.025 GiB, 11.01% gc time, 25.39% compilation time: 4% of which was recompilation)
  1.294125 seconds (12.61 M allocations: 1.006 GiB, 9.30% gc time)
  1.332829 seconds (12.61 M allocations: 1.006 GiB, 10.58% gc time)
  1.315999 seconds (12.61 M allocations: 1.006 GiB, 10.65% gc time)
=====================================
Normal version
  2.476345 seconds (19.59 M allocations: 1.348 GiB, 22.42% gc time, 3.80% compilation time)
  2.391040 seconds (19.51 M allocations: 1.343 GiB, 25.77% gc time)
  2.397999 seconds (19.51 M allocations: 1.343 GiB, 20.28% gc time)
  2.250012 seconds (19.51 M allocations: 1.343 GiB, 20.42% gc time)
  ```
  3. I added `@info` instead of `println` for showing the levels such that we can easily disable the logging when we do not need it.
  4. I added an extra print to show how big the `eval_cache` is.
  5. I added a test that checks that the sum_based iteration finds the same lists as the _bruteforce_ option. For that I generated 10 random inputs and tested that the output is the same.