@nicolaefilat 2024-05-11
@eErr0Re Do you have any other optimizations in mind for the enumerator?
Also, should we keep both the previous option and the optimized version around?