@eErr0Re 2024-05-13
>  Also, should we keep both the previous option and the optimized version around?

I think it is better to just have the optimised version.